fix: fix commented issues, add timeUtil and fileUtil to work with external time and files logic
