+++ frontend/src/pages/server/components/message/MessageComponent.tsx
@@ -1,4 +1,5 @@
 import React, { FunctionComponent, useRef } from 'react'
+import { getDateString, getTimeString } from 'utils/timeUtil'
 
 interface MessageComponentProps {
     content?: string
@@ -24,29 +25,9 @@ const MessageComponent: FunctionComponent<MessageComponentProps> = ({
     // So we need to manually cast to number
     const serverTime = timestamp as unknown as number
 
-    // Get time from timestamp
-    const getMessageTime = () =>
-        new Date(serverTime).toLocaleString([], {
-            hour: '2-digit',
-            minute: '2-digit',
-        })
-
-    const getMessageDateString = () => {
-        const currentTime = new Date()
-        const messageTime = new Date(serverTime)
-
-        const diffDay = currentTime.getDate() - messageTime.getDate()
-        const diffMonth = currentTime.getMonth() - messageTime.getMonth()
-        const diffYear = currentTime.getFullYear() - messageTime.getFullYear()
-
-        if (diffDay === 0 && diffMonth === 0 && diffYear === 0) {
-            return 'Today at '
-        } else if (diffDay === 1 && diffMonth === 0 && diffYear === 0) {
-            return 'Yesterday at '
-        } else {
-            return `${messageTime.getDate()}/${messageTime.getMonth()}/${messageTime.getFullYear()} at `
-        }
-    }
+    const messageTime = `${getDateString(serverTime)}${getTimeString(
+        serverTime,
+    )}`
 
     const onImageLoad = () => {
         onMessageLoaded()
@@ -65,7 +46,7 @@ const MessageComponent: FunctionComponent<MessageComponentProps> = ({
             <div className="ml-4">
                 <div className="flex items-baseline pb-1">
                     <h3 className="font-bold">{createdBy.username}</h3>
-                    <h5 className="text-slack-text-blur px-2">{`${getMessageDateString()}${getMessageTime()}`}</h5>
+                    <h5 className="text-slack-text-blur px-2">{messageTime}</h5>
                 </div>
                 {content && <h5 className="text-lg">{content}</h5>}
                 {media && (
