+++ frontend/src/styles/global.scss
@@ -59,3 +59,7 @@ i .icon {
   // Equal to mt-2
   margin: 0.5rem 0 0 0 !important
 }
+
+#message-panel__content {
+  max-height: 70vh;
+}
