+++ frontend/src/utils/fileUtil.ts
@@ -1,0 +1,4 @@
+export function extractFileExt(fileName: string): string {
+  const lastDotPos = fileName.lastIndexOf('.')
+  return fileName.slice(lastDotPos)
+}
